+++ app/models/suggested_edit.rb
@@ -1,5 +1,6 @@
 class SuggestedEdit < ApplicationRecord
   include PostRelated
+  include PostValidations
 
   belongs_to :user
 
@@ -9,6 +10,9 @@ class SuggestedEdit < ApplicationRecord
   belongs_to :decided_by, class_name: 'User', optional: true
   has_and_belongs_to_many :tags
   has_and_belongs_to_many :before_tags, class_name: 'Tag', join_table: 'suggested_edits_before_tags'
+
+  has_one :post_type, through: :post
+  has_one :category, through: :post
 
   after_save :clear_pending_cache, if: Proc.new { saved_change_to_attribute?(:active) }
